fixing map's coordinates

+++ mandatory/includes/cub.h
@@ -6,7 +6,7 @@
 /*   By: zait-bel <zait-bel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/13 12:09:37 by zait-bel          #+#    #+#             */
-/*   Updated: 2024/10/13 16:01:43 by zait-bel         ###   ########.fr       */
+/*   Updated: 2024/10/13 16:40:12 by zait-bel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -129,7 +129,7 @@ void	handle_key_input(mlx_key_data_t keydata, void* param);
 t_inter	find_vertical_intersection(t_cube *cube, double angle);
 t_inter	find_horizontal_intersection(t_cube *cube, double angle);
 void	initialize_position(t_data *data, double *x, double *y);
-bool	is_wall(int x, int y, t_cube *cub);
+bool	is_wall(size_t x, int y, t_cube *cub);
 void	render_wall(t_cube *cube, double x, double ray);
 void render_minimap(void* param);
 void bresenham_line(long from_x, long from_y, long to_x, long to_y, t_cube *cub, uint32_t color);
